Fix `root_dir` path for different active LSP clients
`root_dir` using the same path in the current working directory for different
active LSP clients.

Example: The current working directory is in `~/.config/nvim`, but if open a
buffer from different LSP client such as `index.html` in `~/project/web` that
use HTML LSP, it will use the same `~/.config/nvim` in client's `root_dir` `config`.

So to solve the problem is to check if the buffer `filetypes` and `root_dir`
are the same will reuse the LSP client. If one of those condition false, it
will create a new LSP client instance.

This changes also fix `root_dir` with a `nil` value if open a buffer that
doesn't find the `root_pattern`, it will set to the current buffer working
directory. I think its like a "single file mode" that nvim-lspconfig do, but I
set to the current buffer working directory.